add TargOrgService

+++ src/main/java/com/terminalmock/test/repositories/enumsrepo/AttrTypeRepo.java
@@ -3,5 +3,5 @@ package com.terminalmock.test.repositories.enumsrepo;
 import com.terminalmock.test.entities.enums.AttrType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttrTypeRepo extends JpaRepository<AttrType, Long> {
+public interface AttrTypeRepo extends JpaRepository<AttrType, Integer> {
 }
